Separate form and list into two components
